move the NextProtos test into a separate file so that it can be run conditionally on go versions >= 1.14
